Compatibility with ServerApp backend (#65)
Jupyter provides nbclassic mode as the compatible notebook backend that uses the jupyter sever backend. It will be fully replaced by the notebookv7 when it is released. But for the moment, if the user what to use the nbclassic mode the appmode is not working. 

This commit makes it compatible with the nbclassic backend by fixing the concurrent issue mentioned in https://discourse.jupyter.org/t/jupyter-notebook-zmq-message-arrived-on-closed-channel-error/17869. It also adds options to allow controlling the path of template files of opened notebooks. 

Since the jupyter server backends don't have the templates needed by the appmode, this commit explicitly adds the template path so those can be found and used.